pyproject.toml
[tool.poetry]
name = "atlas-backend"
version = "0.1.0"
description = ""
authors = ["David Cramer <dcramer@gmail.com>"]
license = "Apache-2.0"
packages = [
    { include = "backend" },
]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.1"
psycopg2-binary = "^2.7"
itsdangerous = "^1.1"
facebook-sdk = "^3.1"
graphene-django = "^2.2"
honcho = "^1.0"
gunicorn = "^19.9"
django-heroku = "^0.3.1"
factory-boy = "^2.11"
lxml = "^4.2"
django-enumfields = "^1.0"
pillow = "^5.3"
google-auth = "^1.6"
sentry-sdk = "^0.8.1"

[tool.poetry.dev-dependencies]
isort = "==4.3.4"
black = "=19.3b0"
flake8 = "^3.6"
pre-commit = "^1.13"
pytest = "^4.0"
pytest-django = "^3.4"
pytest-xdist = "^1.25"
pytest-timeout = "^1.3"

[tool.poetry.scripts]
atlas = 'backend.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
